feat: separate defines
