+++ src/commands/exportApi.ts
@@ -5,8 +5,8 @@ import * as fs from "fs-extra";
 import fetch from 'node-fetch';
 import * as path from "path";
 import * as vscode from "vscode";
-import { ApiCenterVersionDefinitionManagement } from "../azure/ApiCenter/ApiCenterDefinition";
 import { ApiSpecExportResultFormat } from "../azure/ApiCenter/contracts";
+import { ApiCenterVersionDefinitionManagement } from "../azure/ApiCenterDefines/ApiCenterDefinition";
 import { ext } from "../extensionVariables";
 import { ApiVersionDefinitionTreeItem } from "../tree/ApiVersionDefinitionTreeItem";
 import { createTemporaryFolder } from "../utils/fsUtil";
