+++ src/commands/registerApiSubCommands/registerStepByStep.ts
@@ -4,9 +4,9 @@ import { getResourceGroupFromId } from "@microsoft/vscode-azext-azureutils";
 import { IActionContext } from "@microsoft/vscode-azext-utils";
 import * as fse from 'fs-extra';
 import * as vscode from 'vscode';
-import { ApiCenterVersionDefinitionManagement } from "../../azure/ApiCenter/ApiCenterDefinition";
 import { ApiCenterService } from "../../azure/ApiCenter/ApiCenterService";
 import { ApiCenterApi, ApiCenterApiVersion, ApiCenterApiVersionDefinition, ApiCenterApiVersionDefinitionImport, ApiKind, ApiVersionLifecycleStage, SpecificationName } from "../../azure/ApiCenter/contracts";
+import { ApiCenterVersionDefinitionManagement } from "../../azure/ApiCenterDefines/ApiCenterDefinition";
 import { ext } from "../../extensionVariables";
 import { ApiCenterTreeItem } from "../../tree/ApiCenterTreeItem";
 import { ApisTreeItem } from "../../tree/ApisTreeItem";
