+++ src/tree/ApiTreeItem.ts
@@ -2,7 +2,7 @@
 // Licensed under the MIT license.
 import { AzExtParentTreeItem, AzExtTreeItem, IActionContext, TreeItemIconPath } from "@microsoft/vscode-azext-utils";
 import * as vscode from 'vscode';
-import { ApiCenterApiManagement, IApiCenterBase } from "../azure/ApiCenter/ApiCenterDefinition";
+import { ApiCenterApiManagement, IApiCenterApiBase } from "../azure/ApiCenterDefines/ApiCenterApi";
 import { UiStrings } from "../uiStrings";
 import { ApiDeploymentsTreeItem } from "./ApiDeploymentsTreeItem";
 import { ApiVersionsTreeItem } from "./ApiVersionsTreeItem";
@@ -12,13 +12,13 @@ export class ApiTreeItem extends AzExtParentTreeItem {
   public readonly contextValue: string = ApiTreeItem.contextValue;
   public readonly apiVersionsTreeItem: ApiVersionsTreeItem;
   public readonly apiDeploymentsTreeItem?: ApiDeploymentsTreeItem;
-  private readonly _apiCenterApi: IApiCenterBase;
+  private readonly _apiCenterApi: IApiCenterApiBase;
   private readonly _apiCenterName: string;
-  constructor(parent: AzExtParentTreeItem, apiCenterName: string, apiCenterApi: IApiCenterBase) {
+  constructor(parent: AzExtParentTreeItem, apiCenterName: string, apiCenterApi: IApiCenterApiBase) {
     super(parent);
     this._apiCenterName = apiCenterName;
     this._apiCenterApi = apiCenterApi;
-    this.apiVersionsTreeItem = new ApiVersionsTreeItem(this, apiCenterName, apiCenterApi);
+    this.apiVersionsTreeItem = new ApiVersionsTreeItem(this, apiCenterName, apiCenterApi.generateChild());
     if (apiCenterApi instanceof ApiCenterApiManagement) {
       this.apiDeploymentsTreeItem = new ApiDeploymentsTreeItem(this, apiCenterName, (apiCenterApi as ApiCenterApiManagement).getData());
     }
